[FRONT-749]Components export and clean up to use them in editor (#29)
The purpose of this PR is to export and tune components to be used in the editor. Also contains changes to clean up a bit our components by removing comments in over component displayName, default props, etc. Had to remove some helpers that are not needed any more in the UI-kit, were part of application-react