[GRO-3575]Add active state to the button (#432)
Add isActive prop to the button, which causes the button to have different styles depending on its variant. It is used only with transparent variant for now (for Semrush navbar).

In next PR activeColor prop will be changed to enabledColor, as it is misleading - this color isn't applied when isActive is true but when button isn't disabled.