[GRO-5004]Add one side padding to Alert message (#579)
This PR adds prop oneSidePadding to Alert. It causes padding on alert message to be removed from right side as it broke design in new pricing page. This padding is not necessary at all, as it probably was added to make space between icon and message, but was added to both sides. Although it isn't certain that removing it completely won't break any other designs so it is added as new prop.